Create common abstraction class for simple providers
